backend/pyproject.toml
[project]
name = "nutri-agent-backend"
version = "0.1.0"
description = "NutriAgent backend API and meal-plan orchestrator"
readme = "../README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "uvicorn[standard]>=0.42.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "asyncpg>=0.31.0",
    "alembic>=1.18.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pgvector>=0.4.2",
    "celery[redis]>=5.6.3",
    "redis>=6.4.0",
    "loguru>=0.7.3",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic-ai>=1.75.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "pyyaml>=6.0.0",
    "bcrypt>=4.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.0",
]

[project.scripts]
nutri-cli = "cli:main"

# ── pytest ──────────────────────────────────────────
[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"

# ── ruff ────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".venv",
    "alembic/versions",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    # "TCH",  # flake8-type-checking (enable later for optimization)
    "RUF",    # ruff-specific
    "ASYNC",  # flake8-async
    "S",      # flake8-bandit (security)
    "T20",    # flake8-print
]
ignore = [
    "E501",   # line too long (formatter handles it)
    "S101",   # assert in tests is fine
    "S104",   # binding to 0.0.0.0 is intentional
    "S105",   # hardcoded passwords in config defaults are fine for dev
    "S106",   # same
    "B008",   # function call in default arg (FastAPI Depends)
    "UP007",  # X | Y union syntax — keep Optional for compat
    "RUF001", # ambiguous unicode (cyrillic in strings is intentional)
    "RUF002", # ambiguous unicode in docstrings (cyrillic is intentional)
    "RUF003", # ambiguous unicode in comments (cyrillic is intentional)
    "UP042",  # str+Enum → StrEnum (SQLAlchemy Enum column requires str+Enum)
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["F401", "S101", "T20"]
"scripts/*.py" = ["T20", "ASYNC"]
"cli.py" = ["T20", "ASYNC"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
